Refactor: rename todaySection
