+++ src/components/SectionIdea/Search.tsx
@@ -1,0 +1,3 @@
+export const Search = () => {
+  return <div>Search</div>;
+};
