+++ src/components/SectionIdea/index.tsx
@@ -1,0 +1,1 @@
+export { Search } from "./Search";
