+++ src/components/SectionToday/TodayIdea.tsx
