--- src/pages/IdeaPage/IdeaSearch.tsx
@@ -1,11 +1,0 @@
-import IdeaBar from "../../components/Sidebar/IdeaBar";
-
-function IdeaSearch() {
-  return (
-    <section>
-      <IdeaBar />
-    </section>
-  );
-}
-
-export default IdeaSearch;
