+++ src/pages/IdeaPage/IdeaSearchPage.tsx
@@ -1,0 +1,13 @@
+import { Search } from "../../components/SectionIdea/Search";
+import { IdeaBar } from "../../components/Sidebar/IdeaBar";
+
+function IdeaSearchPage() {
+  return (
+    <section>
+      <IdeaBar />
+      <Search />
+    </section>
+  );
+}
+
+export default IdeaSearchPage;
